🖌️ Modify : 타입 파일 분리
작업 내용 :
1. ReactNavigation 타입 파일 추가 (전역 타입을 추가하기 위해 ESLint Rule 일부 off)
2. CardModal 타입 추가
